requirements.txt
ansible-core==2.15.6
ansible-cmdb==1.31
google-auth==2.21.0
google-auth-oauthlib==0.7.1
google-auth-httplib2==0.1.0
google-api-python-client==2.96.0
oauth2client==4.1.3
mysql-connector-python==8.0.33
